pyproject.toml
[project]
name = "mastr-mcp-server"
version = "1.0.0"
authors = [
    { name = "mastr-mcp-server contributors" },
]
description = "MCP Server for the German Marktstammdatenregister (MaStR) — energy market register of the Bundesnetzagentur"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["mcp", "mastr", "marktstammdatenregister", "energy", "bundesnetzagentur", "claude"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "mcp>=1.26.0",
    "zeep>=4.2.1",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
mastr-mcp-server = "mastr_mcp_server:mcp.run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
